requirements_enhanced.txt
# Core crawling libraries
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
trafilatura==1.11.0
markdownify==0.12.1

# LLM orchestration
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
pydantic==2.9.2

# URL handling and normalization
tldextract==5.1.2
url-normalize==1.4.3
robotexclusionrulesparser==1.7.1
python-slugify==8.0.4

# Text processing
ftfy==6.2.3
langdetect==1.0.9

# Database and caching
requests-cache==1.2.1

# Reliability and retries
tenacity==9.0.0

# Monitoring and visualization
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3
numpy==2.1.2

# Environment management
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-httpserver==1.1.0
pytest-asyncio==0.24.0

# Progress bars
tqdm==4.66.5

# Utilities
colorama==0.4.6
